requirements.txt
beautifulsoup4
python-dotenv
requests
traktexport